# docs/contracts/BASIC_CURRENT_DEFINITION_OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DCP-BASIC-NO-AUDIT-VERSION-001 bounded API supplement",
    "version": "2026-10-09",
    "description": "Approved supplement to FINAL BASELINE v1.0.23. Only listed basic controls/process bindings supersede prior contracts. Historical evidence and regulated controls remain. Fully qualified schema IDs prevent Create/Update collisions without changing HTTP payload fields. DCP-EBR-TEMPLATE-NAME-001 adds required create templateName and controlled draft name editing; historical canonical content remains unchanged."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/api/v1/units/{id}": {
      "get": {
        "tags": [
          "unit-controller"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "put": {
        "tags": [
          "unit-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.UnitCommands.Update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
    "/api/v1/unit-conversions/{id}": {
      "get": {
        "tags": [
          "unit-conversion-controller"
        ],
        "operationId": "get_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "put": {
        "tags": [
          "unit-conversion-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.UnitConversionCommands.Update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
...
    "/api/v1/units": {
      "get": {
        "tags": [
          "unit-controller"
        ],
        "operationId": "list_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arg3",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "post": {
        "tags": [
          "unit-controller"
        ],
        "operationId": "create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.UnitCommands.Create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
    "/api/v1/unit-conversions": {
      "get": {
        "tags": [
          "unit-conversion-controller"
        ],
        "operationId": "list_2",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arg3",
            "in": "query",
            "required": true,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      },
      "post": {
        "tags": [
          "unit-conversion-controller"
        ],
        "operationId": "create_1",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.UnitConversionCommands.Create"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode"
                }
              }
            }
          }
        },
        "x-basic-maintenance-contract": "No audit append or revision/client-version precondition. Permission, org scope, validation, row locks and idempotency retained."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "com.hospital.mes.common.api.ApiResponseCom.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/com.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "com.hospital.mes.masterdata.application.UnitConversionCommands.Update": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "fromUnitId": {
            "type": "string"
          },
          "toUnitId": {
            "type": "string"
          },
          "factor": {
            "type": "string"
          },
          "materialId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "com.hospital.mes.common.api.ApiResponseCom.fasterxml.jackson.databind.JsonNode": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/com.fasterxml.jackson.databind.JsonNode"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "com.fasterxml.jackson.databind.JsonNode": {},
...
      "com.hospital.mes.masterdata.application.UnitConversionCommands.Create": {
        "type": "object",
        "properties": {
          "fromUnitId": {
            "type": "string"
          },
          "toUnitId": {
            "type": "string"
          },
          "factor": {
            "type": "string"
          },
          "materialId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "com.hospital.mes.masterdata.application.UnitCommands.Create": {
        "type": "object",
        "properties": {
          "unitCode": {
            "type": "string"
          },
          "unitName": {
            "type": "string"
          },
          "dimension": {
            "type": "string"
          },
          "scale": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "com.hospital.mes.masterdata.application.UnitCommands.Update": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "unitName": {
            "type": "string"
          },
          "dimension": {
            "type": "string"
          },
          "scale": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "com.hospital.mes.masterdata.application.ScopedStore.PageDataCom.fasterxml.jackson.databind.JsonNode": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/com.fasterxml.jackson.databind.JsonNode"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}